added login feature
added login feature, but there is an issue, which is that even after registring or logging in, welcome screen is not shown. i dont see error messages regarding this, and i dont know where to correct. i know it is functioning because i can see so in the console and also firebase.